Fix only last call to SVGDoc.draw() working
Clearing the buffers on each call means that only the last call has
an effect, which prevents drawing multiple SVGs (or even the same one
multiple times) in one draw cycle.

Clearing should be (and in the case of the examples, is) done by the
calling application (eg., Pyglet's window.clear() during the on_draw
event).

(Fixes issue #3)